requirements.txt

Flask==2.3.3
Werkzeug==2.3.7
pandas==2.0.3
scikit-learn==1.3.0
spacy==3.6.1
pdfplumber==0.10.2
python-docx==0.8.11
numpy==1.24.3